--- app/services/matchers/fowlerwelch/model2/matcher.js
@@ -1,59 +1,0 @@
-const MatcherResult = require("../../../matches-result");
-const FileExtension = require("../../../../utilities/file-extension");
-
-function matches(packingList, filename) {
-  let headerRow = 0;
-
-  try {
-    // check for correct extension
-
-    if (FileExtension.matches(filename, "xlsx") !== MatcherResult.CORRECT) {
-      return MatcherResult.WRONG_EXTENSIONS;
-    }
-
-    //check for correct establishment number
-
-    const sheets = Object.keys(packingList);
-
-    if (sheets.length === 0) {
-      throw new Error("generic error");
-    }
-
-    for (let i = 0; i < sheets.length; i++) {
-      headerRow = packingList[sheets[i]].findIndex(
-        (x) => x.F === "Description of goods",
-      );
-      const establishmentNumber = packingList[sheets[i]][headerRow + 1].M;
-      const regex = /^RMS-GB-000216-\d{3}$/;
-      if (!regex.test(establishmentNumber)) {
-        return MatcherResult.WRONG_ESTABLISHMENT_NUMBER;
-      }
-
-      // check for header values
-
-      const header = {
-        C: "Commodity code",
-        F: "Description of goods",
-        H: "No. of pkgs",
-        K: "Item Net Weight (kgs)",
-        N: "Treatment Type (Chilled /Ambient)",
-      };
-
-      for (const key in header) {
-        if (
-          (key === "K" &&
-            !packingList[sheets[i]][headerRow][key].includes("Net Weight")) ||
-          (key !== "K" &&
-            !packingList[sheets[i]][headerRow][key].startsWith(header[key]))
-        ) {
-          return MatcherResult.WRONG_HEADER;
-        }
-      }
-    }
-    return MatcherResult.CORRECT;
-  } catch (err) {
-    return MatcherResult.GENERIC_ERROR;
-  }
-}
-
-module.exports = { matches };
