+++ test/unit/services/matchers/fowlerwelch/model1/matcher.test.js
@@ -1,28 +1,22 @@
-const Matcher = require("../../../../../../app/services/matchers/fowlerwelch/model1/matcher");
 const MatcherResult = require("../../../../../../app/services/matches-result");
+const fowlerWelchMatcher = require("../../../../../../app/services/matchers/fowlerwelch/model1/matcher");
 
-describe("matchesFowlerWelchModel1", () => {
+describe("matchesFowlerWelch2", () => {
   test("returns generic error for empty json", () => {
     const packingListJson = {};
     const filename = "packinglist.xlsx";
-
-    const result = Matcher.matches(packingListJson, filename);
-
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
     expect(result).toBe(MatcherResult.GENERIC_ERROR);
   });
-
   test("returns wrong extension for incorrect file extension", () => {
     const filename = "packinglist.xls";
     const packingListJson = {};
-
-    const result = Matcher.matches(packingListJson, filename);
-
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
     expect(result).toBe(MatcherResult.WRONG_EXTENSIONS);
   });
-
-  test("returns wrong establishment number for missing establishment number", () => {
+  test("returns wrong establishment number for missing establishment number for one sheet", () => {
     const packingListJson = {
-      "Customer Order": [
+      "Cust Ord - Vitacress": [
         {},
         {},
         {},
@@ -73,17 +67,121 @@ describe("matchesFowlerWelchModel1", () => {
         },
       ],
     };
-    const filename = "packinglist-wrong-estblishment.xlsx";
-
-    const result = Matcher.matches(packingListJson, filename);
-
+    const filename = "packinglist.xlsx";
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
     expect(result).toBe(MatcherResult.WRONG_ESTABLISHMENT_NUMBER);
   });
-
-  test("returns wrong header for incorrect header values", () => {
+  test("returns wrong establishment number for missing establishment numbers of multiple sheets", () => {
+    const packingListJson = {
+      "Cust Ord - Vitacress": [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          M: "Incorrect",
+        },
+      ],
+      ARGO: [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          M: "RMS-GB-000216-001",
+        },
+      ],
+    };
+    const filename = "packinglist.xlsx";
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
+    expect(result).toBe(MatcherResult.WRONG_ESTABLISHMENT_NUMBER);
+  });
+  test("returns wrong header for incorrect header values of one sheet", () => {
     const filename = "packinglist.xlsx";
     const packingListJson = {
-      "Customer Order": [
+      "Cust Ord - Vitacress": [
         {},
         {},
         {},
@@ -139,30 +237,24 @@ describe("matchesFowlerWelchModel1", () => {
           H: "No. of pkgs",
           I: "Type of pkgs",
           J: "Total Gross Weight",
-          K: "Total Net Weight",
-          L: "Total Line Value",
+          K: "Item Net Weight (kgs)",
+          L: "Item Value",
           M: "NIIRMS Dispatch number",
           N: "Treatment Type (Chilled /Ambient)",
           O: "NIRMS Lane (R/G)",
-          P: "Secondary Qty",
-          Q: "Cert Type Req",
-          R: "Cert Number",
         },
         {
-          M: "RMS-GB-000216-004",
+          M: "RMS-GB-000216-002",
         },
       ],
     };
-
-    const result = Matcher.matches(packingListJson, filename);
-
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
     expect(result).toBe(MatcherResult.WRONG_HEADER);
   });
-
-  test("returns Correct", () => {
+  test("returns wrong header for incorrect header values of multiple sheets", () => {
     const filename = "packinglist.xlsx";
     const packingListJson = {
-      "Customer Order": [
+      "Cust Ord - Vitacress": [
         {},
         {},
         {},
@@ -208,33 +300,392 @@ describe("matchesFowlerWelchModel1", () => {
         {},
         {},
         {
-          A: "Item",
-          B: "Product code",
-          C: "Commodity code",
+          A: "NOT",
+          B: "CORRECT",
+          C: "HEADER",
           D: "Online Check",
-          E: "Meursing code",
+          E: "Meursing Code",
           F: "Description of goods",
           G: "Country of Origin",
-          H: "No. of pkgs \r\n(1547)",
+          H: "No. of pkgs",
           I: "Type of pkgs",
-          J: "Total Gross Weight \r\n(11015.700kgs)",
-          K: "Total Net Weight \r\n(7921.700kgs)",
-          L: "Total Line Value \r\n(41662.4)",
+          J: "Total Gross Weight",
+          K: "Item Net Weight (kgs)",
+          L: "Item Value",
           M: "NIIRMS Dispatch number",
           N: "Treatment Type (Chilled /Ambient)",
           O: "NIRMS Lane (R/G)",
-          P: "Secondary Qty",
-          Q: "Cert Type Req",
-          R: "Cert Number",
         },
         {
-          M: "RMS-GB-000216-004",
+          M: "RMS-GB-000216-002",
+        },
+      ],
+      ARGO: [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "No. of pkgs",
+          K: "Item Net Weight (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-001",
         },
       ],
     };
-
-    const result = Matcher.matches(packingListJson, filename);
-
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
+    expect(result).toBe(MatcherResult.WRONG_HEADER);
+  });
+  test("returns correct for correct headers for one sheet", () => {
+    const filename = "packinglist.xlsx";
+    const packingListJson = {
+      "Cust Ord - Vitacress": [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "No. of pkgs",
+          K: "Item Net Weight (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-002",
+        },
+      ],
+    };
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
     expect(result).toBe(MatcherResult.CORRECT);
+  });
+  test("returns correct for correct headers of multiple sheets", () => {
+    const filename = "packinglist.xlsx";
+    const packingListJson = {
+      "Cust Ord - Vitacress": [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "No. of pkgs",
+          K: "Item Net Weight (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-002",
+        },
+      ],
+      ARGO: [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "No. of pkgs",
+          K: "Item Net Weight (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-001",
+        },
+      ],
+    };
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
+    expect(result).toBe(MatcherResult.CORRECT);
+  });
+  test("if the key is equal to 'K' and doesn't include 'Net Weight' in its header, return wrong header", () => {
+    const filename = "packinglist.xlsx";
+    const packingListJson = {
+      "Cust Ord - Vitacress": [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "No. of pkgs",
+          K: "Item (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-002",
+        },
+      ],
+    };
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
+    expect(result).toBe(MatcherResult.WRONG_HEADER);
+  });
+  test("if the header doesn't start with the header[key], return wrong header", () => {
+    const filename = "packinglist.xlsx";
+    const packingListJson = {
+      "Cust Ord - Vitacress": [
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {},
+        {
+          C: "Commodity code",
+          F: "Description of goods",
+          H: "(318)",
+          K: "Item Net Weight (kgs)",
+          N: "Treatment Type (Chilled /Ambient)",
+        },
+        {
+          M: "RMS-GB-000216-002",
+        },
+      ],
+    };
+    const result = fowlerWelchMatcher.matches(packingListJson, filename);
+    expect(result).toBe(MatcherResult.WRONG_HEADER);
   });
 });
